List content of the CLD

Returns a named list of legislatures and sessions available in the CLD. This provides a quick overview of the CLD's scope and valid three-letter country codes, and helps to conventiently loop/map over legislatures and sessions.

cld_content(legislature = NULL)

Arguments

legislature
An optional character string specifying one or more legislatures. Currently one of ‘aut’, ‘can’, ‘cze’, ‘esp’, ‘fra’, ‘deu’, ‘irl’, ‘sco’, ‘gbr’, ‘usa_house’, or ‘usa_senate’. If NULL (the default), a list with all legislatures and sessions available in the CLD is returned.

Value

A list with names being three-letter country codes and with each element containing a vector that shows the sessions available for a legislature.

Examples

# Get a list of three-letter country codes and available sessions for all countries
overview <- cld_content()
tibble::glimpse(overview)

#> List of 11
#>  $ aut       : int [1:27] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ can       : int [1:43] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ cze       : int [1:8] 1 2 3 4 5 6 7 8
#>  $ esp       : int [1:15] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ fra       : int [1:15] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ deu       : int [1:19] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ irl       : int [1:33] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ sco       : int [1:5] 1 2 3 4 5
#>  $ gbr       : int [1:58] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ usa_house : int [1:116] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ usa_senate: int [1:116] 1 2 3 4 5 6 7 8 9 10 ...

# Get a list of available sessions for the French Assemblée and the Irish Dail
sessions <- cld_content(legislature = c("fra", "irl"))
tibble::glimpse(sessions)

#> List of 2
#>  $ fra: int [1:15] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ irl: int [1:33] 1 2 3 4 5 6 7 8 9 10 ...