Fetch 'Core' table

Fetches sociodemographic data of legislators for the specified legislature. Requires a working Internet connection.

get_core(legislature)

Arguments

legislature
A character string specifying the three-letter country code of the legislature for which data shall be fetched. Currently one of ‘aut’, ‘can’, ‘cze’, ‘esp’, ‘fra’, ‘deu’, ‘irl’, ‘sco’, ‘gbr’, ‘usa_house’, or ‘usa_senate’.

Format

Data frame with columns (varies by legislature):

country: ISO 3166-1 alpha-3 three-letter country code (of class ‘character’).

pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class ‘integer’ or ‘character’).

wikidataid: Wikidata ID identifying a legislator's Wikidata entry (of class ‘character’).

wikititle: A legislator's undirected Wikipedia title (of class ‘character’).

name: A legislator's full name (of class ‘character’).

sex: A legislator's sex (of class ‘character’).

ethnicity: A legislator's ethnicity (of class ‘character’).

religion: A legislator's religious denomination (of class ‘character’).

birth: A legislator's date of birth (of class ‘POSIXct’).

death: A legislator's date of death (of class ‘POSIXct’).

birthplace: Comma separated latitude and longitude of a legislator's place of birth (of class ‘character’).

deathplace: Comma separated latitude and longitude of a legislator's place of death (of class ‘character’).

Source

Wikipedia, https://wikipedia.org/
Wikipedia API, https://wikipedia.org/w/api.php
Wikidata API, https://www.wikidata.org/
Wikimedia Commons, https://commons.wikimedia.org/
Face++ Cognitive Services API, https://www.faceplusplus.com/
Germany Bundestag Open Data, https://www.bundestag.de/service/opendata

Value

A data frame with columns as specified above.

Examples

# Get entire 'Core' table for the German Bundestag
deu_core <- get_core(legislature = "deu")
tibble::glimpse(deu_core)

#> Rows: 4,075
#> Columns: 12
#> $ country    <chr> "DEU", "DEU", "DEU", "DEU", "DEU", "DEU", "DEU", "DEU", "D…
#> $ pageid     <chr> "2657", "171955", "171957", "171967", "172061", "172064", …
#> $ wikidataid <chr> "Q2492", "Q100613", "Q360721", "Q1382213", "Q1277445", "Q1…
#> $ wikititle  <chr> "Konrad_Adenauer", "Willi_Agatz", "Adolf_Ahrens_(Kapitän)"…
#> $ name       <chr> "Konrad Adenauer", "Willi Agatz", "Adolf Ahrens", "Johanne…
#> $ sex        <chr> "male", "male", "male", "male", "female", "female", "male"…
#> $ ethnicity  <chr> "white", NA, NA, NA, NA, NA, NA, "white", "white", NA, "wh…
#> $ religion   <chr> "catholicism", NA, "protestantism evangelical", "catholici…
#> $ birth      <dttm> 1876-01-05, 1904-01-01, 1879-09-17, 1890-03-08, 1901-06-2…
#> $ death      <dttm> 1967-04-19, 1957-01-01, 1957-01-21, 1963-03-08, 1979-02-0…
#> $ birthplace <chr> "50.94222,6.95778", "51.45083,7.01306", "53.5364,8.60056",…
#> $ deathplace <chr> "50.6595,7.21282", "52.51667,13.38333", "53.07694,8.80889"…

# Get 'Core' table for 16th session of the German Bundestag
deu_core_subset <- dplyr::semi_join(x = deu_core,
                                    y = dplyr::filter(get_political(legislature = "deu"),
                                                      session == 16),
                                    by = "pageid")
tibble::glimpse(deu_core_subset)

#> Rows: 641
#> Columns: 12
#> $ country    <chr> "DEU", "DEU", "DEU", "DEU", "DEU", "DEU", "DEU", "DEU", "D…
#> $ pageid     <chr> "12609", "1596", "2655362", "111229", "174245", "175509", …
#> $ wikidataid <chr> "Q65394", "Q58609", "Q16019", "Q61803", "Q1870866", "Q1289…
#> $ wikititle  <chr> "Herta_Däubler-Gmelin", "Franz_Müntefering", "Wolfgang_Sch…
#> $ name       <chr> "Herta Däubler-Gmelin", "Franz Müntefering", "Wolfgang Sch…
#> $ sex        <chr> "female", "male", "male", "male", "male", "male", "male", …
#> $ ethnicity  <chr> "white", "white", "white", "white", NA, "white", "white", …
#> $ religion   <chr> "protestantism evangelical", "catholicism", "protestantism…
#> $ birth      <dttm> 1943-08-12, 1940-01-16, 1942-09-18, 1944-12-14, 1944-12-2…
#> $ death      <dttm> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ birthplace <chr> "48.14472,17.11278", "51.42861,7.99611", "47.99472,7.84972…
#> $ deathplace <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…