Fetch Wikipedia 'History' table

Fetches full revision histories of legislators' Wikipedia biographies for the specified legislature. Requires a working Internet connection.

get_history(legislature)

Arguments

legislature
A character string specifying the three-letter country code of the legislature for which data shall be fetched. Currently one of ‘aut’, ‘can’, ‘cze’, ‘esp’, ‘fra’, ‘deu’, ‘irl’, ‘sco’, ‘gbr’, ‘usa_house’, or ‘usa_senate’.

Format

Data frame with columns:

pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class ‘integer’).

revid: Wikipedia edit ID (of class ‘integer’).

parentid: Wikipedia edit ID of the previous revision (of class ‘integer’).

user: Username of registered user responsible for the revision, IP address in case of anonymous revision (of class ‘character’).

userid: ID of registered user responsible for the revision, 0 in case of anonymous revision (of class ‘integer’).

timestamp: Date and time of the revision (of class ‘POSIXct’).

size: Revision size in bytes (of class ‘integer’).

comment: Revision comment (of class ‘character’).

Source

Wikipedia API, https://wikipedia.org/w/api.php

Value

A data frame with columns as specified above.

Examples

# Get entire 'History' table for the Austrian Nationalrat
aut_history <- get_history(legislature = "aut")
tibble::glimpse(aut_history)

#> Rows: 85,049
#> Columns: 8
#> $ pageid    <chr> "3041286", "3041286", "3041286", "3041286", "3041286", "304…
#> $ revid     <int> 153245941, 151971371, 147949177, 147947191, 144633337, 1375…
#> $ parentid  <int> 151971371, 147949177, 147947191, 144633337, 137542895, 1368…
#> $ user      <chr> "Boshomi", "Invisigoth67", "Liberaler Humanist", "Liberaler…
#> $ userid    <int> 1294898, 178175, 138009, 138009, 116228, 41136, 41136, 1486…
#> $ timestamp <dttm> 2016-04-06 21:59:44, 2016-02-27 15:41:18, 2015-11-11 21:35…
#> $ size      <int> 2326, 2329, 2334, 2336, 2308, 2268, 2238, 2251, 2250, 2239,…
#> $ comment   <chr> "Ersetzungen", "/* Politik */ gramm.", "Markierung angepass…

# Get 'History' table for NEOS party members of the Austrian Nationalrat
aut_history_subset <- dplyr::semi_join(x = aut_history,
                                       y = dplyr::filter(get_political(legislature = "aut"),
                                                         party == "NEOS"),
                                       by = "pageid")
tibble::glimpse(aut_history_subset)

#> Rows: 1,139
#> Columns: 8
#> $ pageid    <chr> "6346640", "6346640", "6346640", "6346640", "6346640", "634…
#> $ revid     <int> 183326844, 180714737, 180714381, 174860214, 169608087, 1644…
#> $ parentid  <int> 180714737, 180714381, 174860214, 169608087, 164468527, 1644…
#> $ user      <chr> "InternetArchiveBot", "M2k~dewiki", "2001:62A:4:41B:5581:14…
#> $ userid    <int> 2458679, 45933, 0, 1481078, 886856, 45933, 45933, 34830, 34…
#> $ timestamp <dttm> 2018-12-02 10:41:45, 2018-09-07 14:41:45, 2018-09-07 14:26…
#> $ size      <int> 13709, 13416, 13349, 12815, 12815, 12790, 12794, 12594, 124…
#> $ comment   <chr> "[[:Benutzer:InternetArchiveBot|InternetArchiveBot]] hat 2 …