Fetch 'Political' table

Fetches political data of legislators for the specified legislature. Requires a working Internet connection.

get_political(legislature)

Arguments

legislature
A character string specifying the three-letter country code of the legislature for which data shall be fetched. Currently one of ‘aut’, ‘can’, ‘cze’, ‘esp’, ‘fra’, ‘deu’, ‘irl’, ‘sco’, ‘gbr’, ‘usa_house’, or ‘usa_senate’.

Format

Data frame in long format with columns (varies by legislature):

pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class ‘integer’).

session: Legislative period (of class ‘integer’).

party: A legislator's party affiliation (of class ‘character’). See https://github.com/saschagobel/legislatoR for the full form of abbreviated party names and english translations of non-english party names

constituency: A legislator's constituency (of class ‘character’).

constituency2: A legislator's constituency (upper level, if applicable, of class ‘character’).

constituency_id: ID of a legislator's constituency (of class ‘character’).

session_start: Date the legislative period started (of class ‘Date’).

session_end: Date the legislative period ended (of class ‘Date’).

service: A legislator's period of service in days during the respective session (of class ‘integer’)

government (or similar): Indicator of a legislator's majority status in parliament (of class ‘logical’). Further columns with extensions of this might exist.

leader (or similar): Indicator of a legislator's leader status in parliament (of class ‘logical’). Further columns with extensions of this might exist.

Source

Wikipedia, https://wikipedia.org/
Czech Republic Parliamentary Members Archive http://public.psp.cz/sqw/fsnem.sqw?zvo=1
Spain Parliamentary Members Archive http://www.congreso.es/portal/page/portal/Congreso/Congreso/Diputados

Value

A data frame with columns as specified above.

Examples

# Get entire 'Political' table for the Czech Poslanecka Snemovna
cze_political <- get_political(legislature = "cze")
tibble::glimpse(cze_political)

#> Rows: 1,718
#> Columns: 7
#> $ pageid        <chr> "844662", "844668", "844679", "337896", "179718", "8447…
#> $ session       <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ party         <chr> "ODS", "ODS", "ODS", "KDU-CSL", "KDS", "MNS", "CSSD", "…
#> $ constituency  <chr> "Hlavní mìsto Praha", "Západoèeský", "Západoèeský", "Ji…
#> $ session_start <date> 1992-06-06, 1992-06-06, 1992-06-06, 1992-06-06, 1992-0…
#> $ session_end   <date> 1996-06-01, 1996-06-01, 1996-06-01, 1996-06-01, 1996-0…
#> $ service       <int> 1461, 1461, 1461, 1461, 1461, 1461, 1461, 1461, 1461, 1…

# Get 'Political' table for female DSP party members of the Czech Poslanecka Snemovna
cze_political_subset <- dplyr::semi_join(x = dplyr::filter(cze_political,
                                                           party == "ODS"),
                                         y = dplyr::filter(get_core(legislature = "cze"),
                                                           sex == "female"),
                                         by = "pageid")
tibble::glimpse(cze_political_subset)

#> Rows: 59
#> Columns: 7
#> $ pageid        <chr> "847370", "848027", "178105", "853926", "856132", "8583…
#> $ session       <int> 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3…
#> $ party         <chr> "ODS", "ODS", "ODS", "ODS", "ODS", "ODS", "ODS", "ODS",…
#> $ constituency  <chr> "Severoèeský", "Hlavní mìsto Praha", "Hlavní mìsto Prah…
#> $ session_start <date> 1992-06-06, 1992-06-06, 1992-06-06, 1992-06-06, 1992-0…
#> $ session_end   <date> 1996-06-01, 1996-06-01, 1996-06-01, 1996-06-01, 1996-0…
#> $ service       <int> 1461, 1461, 1461, 1461, 1461, 1461, 748, 748, 748, 748,…