Fetch Wikipedia 'Traffic' table

Fetches daily user traffic on legislators' Wikipedia biographies for the specified legislature. Requires a working Internet connection.

get_traffic(legislature)

Arguments

legislature
A character string specifying the three-letter country code of the legislature for which data shall be fetched. Currently one of ‘aut’, ‘can’, ‘cze’, ‘esp’, ‘fra’, ‘deu’, ‘irl’, ‘sco’, ‘gbr’, ‘usa_house’, or ‘usa_senate’.

Format

Data frame with columns:

pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class ‘integer’).

date: Date for which user traffic is recorded, from 2015-07-01 to 2018-12-31 UTC (of class ‘POSIXct’).

traffic: Daily non-unique user visits (of class ‘numeric’).

Source

Wikimedia API, https://wikimedia.org/api/rest_v1/
http://petermeissner.de:8880/

Value

A data frame with columns as specified above.

Examples

# Get entire 'Traffic' table for the Scottish Parliament
sco_traffic <- get_traffic(legislature = "sco")
tibble::glimpse(sco_traffic)

#> Rows: 1,230,266
#> Columns: 3
#> $ pageid  <chr> "472409", "472409", "472409", "472409", "472409", "472409", "…
#> $ date    <dttm> 2007-12-10, 2007-12-11, 2007-12-12, 2007-12-13, 2007-12-14, …
#> $ traffic <dbl> 7, 9, 11, 10, 15, 5, 12, 7, 12, 21, 12, 8, 15, 8, 13, 14, 8, …

# Add Wikidataid to 'Traffic' table for the Scottish Parliament
sco_traffic_subset <- dplyr::inner_join(x = dplyr::select(get_core(legislature = "sco"),
                                                          pageid, wikidataid),
                                        y = sco_traffic,
                                        by = "pageid")
tibble::glimpse(sco_traffic_subset)

#> Rows: 1,230,266
#> Columns: 4
#> $ pageid     <chr> "472409", "472409", "472409", "472409", "472409", "472409"…
#> $ wikidataid <chr> "Q570298", "Q570298", "Q570298", "Q570298", "Q570298", "Q5…
#> $ date       <dttm> 2007-12-10, 2007-12-11, 2007-12-12, 2007-12-13, 2007-12-1…
#> $ traffic    <dbl> 7, 9, 11, 10, 15, 5, 12, 7, 12, 21, 12, 8, 15, 8, 13, 14, …