legislatoR

Facilitates access to the Comparative Legislators Database (CLD). The CLD includes political, sociodemographic, career, online presence, public attention, and visual information for over 45,000 contemporary and historical politicians from ten countries.

legislatoR functions

cld_content: returns a named list of legislatures and sessions available in the CLD.

get_core: fetches sociodemographic data of legislators.

get_political: fetches political data of legislators.

get_history: fetches full revision histories of legislators' Wikipedia biographies.

get_traffic: fetches daily user traffic on legislators' Wikipedia biographies.

get_social: fetches social media handles and website URLs of legislators.

get_portrait: fetches portrait urls of legislators.

get_office: fetches political and other offices of legislators.

get_profession fetches occupational data of legislators.

get_ids: fetches a range of IDs of legislators.